Use a single menu on the navbar
Until now you would specify which menus to use on the navigation bar,
and these menus would show one beside the other; they'd only have one
level. Now you only use a single menu, which has (up to) two levels, and
if a menu item has subitems it shows as a submenu. Among other things,
this makes it possible to have some buttons (not submenus) in the
navbar.